Cookies are now properly displaying.
Some onload functions (such as to check if there are cookies) will need
to be added to your games pages, but otherwise the cookie values display
properly. You will have to update your game logic to show that the cash
amount updates accordingly when the user places bets and wins/looses.