Slot
I added the game, The animation is working.  There's a little formatting
left to do. I'll do it soon. Also I'll work on my game ligic.  For some
reason I always get a error with the cookies script saying that fname is
not found or someshit.  I had to delete de onload event on the body tag
so it could start.
***All the modifications I did were applied on my game, I didnt change
any other folder**    :P