Money flash functions
Functions allowing the money div above to flash when winning/losing have
now been added to the template javascript so that all games have access
to it. You will need to integrate it in your game yourself, but it is
easy. You simply need to call the respective functions depending on the
situation.